Subiendo modificaciones de estilos css y estructuración html, implementando generación y descarga de pdf para los informes, cambiando a punto flotante los valores de los impuestos

+++ backoffice-Back-End/src/main/java/edu/bootcamp/backoffice/model/Tax/dto/UpdateChargeRequest.java
@@ -16,7 +16,7 @@ public class UpdateChargeRequest {
     private String name;
 
     @JsonProperty("percentage")
-    private Integer percentage;
+    private Float percentage;
 
     @JsonProperty("enabled")
     private Boolean enabled;
